+++ packages/job-worker/src/blueprints/__tests__/context-events.test.ts
@@ -1,17 +1,9 @@
-import * as _ from 'underscore'
-import { IBlueprintSegmentDB, IBlueprintPieceInstance } from '@sofie-automation/blueprints-integration'
-import {
-	PartEventContext,
-	OnTimelineGenerateContext,
-	RundownDataChangedEventContext,
-	RundownTimingEventContext,
-} from '../context'
-import { PartInstanceId, PieceInstanceInfiniteId, PartId } from '@sofie-automation/corelib/dist/dataModel/Ids'
+import { IBlueprintSegmentDB } from '@sofie-automation/blueprints-integration'
+import { PartEventContext, RundownDataChangedEventContext, RundownTimingEventContext } from '../context'
 import { DBPart } from '@sofie-automation/corelib/dist/dataModel/Part'
 import { DBPartInstance } from '@sofie-automation/corelib/dist/dataModel/PartInstance'
 import { PieceInstance } from '@sofie-automation/corelib/dist/dataModel/PieceInstance'
-import { ABSessionInfo, DBRundownPlaylist } from '@sofie-automation/corelib/dist/dataModel/RundownPlaylist'
-import { OnGenerateTimelineObjExt } from '@sofie-automation/corelib/dist/dataModel/Timeline'
+import { DBRundownPlaylist } from '@sofie-automation/corelib/dist/dataModel/RundownPlaylist'
 import { getRandomId } from '@sofie-automation/corelib/dist/lib'
 import { protectString, unprotectString } from '@sofie-automation/corelib/dist/protectedString'
 import { MockJobContext, setupDefaultJobEnvironment } from '../../__mocks__/context'
@@ -369,572 +361,6 @@ describe('Test blueprint api context', () => {
 			const segment = (await context.getSegment(`${rundown._id}_segment2`)) as IBlueprintSegmentDB
 			expect(segment).toBeTruthy()
 			expect(segment._id).toEqual(`${rundown._id}_segment2`)
-		})
-	})
-
-	describe('OnTimelineGenerateContext', () => {
-		const getSessionId = (n: number): string => `session#${n}`
-		async function getContext(
-			rundown: DBRundown,
-			previousPartInstance: DBPartInstance | undefined,
-			currentPartInstance: DBPartInstance | undefined,
-			nextPartInstance: DBPartInstance | undefined
-		) {
-			const playlist = (await jobContext.mockCollections.RundownPlaylists.findOne(
-				rundown.playlistId
-			)) as DBRundownPlaylist
-			expect(playlist).toBeTruthy()
-
-			const context = new OnTimelineGenerateContext(
-				jobContext.studio,
-				jobContext.getStudioBlueprintConfig(),
-				showStyle,
-				jobContext.getShowStyleBlueprintConfig(showStyle),
-				playlist,
-				rundown,
-				previousPartInstance,
-				currentPartInstance,
-				nextPartInstance
-			)
-
-			let nextId = 0
-			context.getNewSessionId = () => getSessionId(nextId++)
-
-			return context
-		}
-
-		function getAllKnownSessions(context: OnTimelineGenerateContext): ABSessionInfo[] {
-			const sessions: ABSessionInfo[] = (context as any)._knownSessions
-			expect(sessions).toBeTruthy()
-
-			return sessions.map((s) => _.omit<ABSessionInfo & { keep?: boolean }, 'keep'>(s, 'keep'))
-		}
-		// function overwriteKnownSessions(context: TimelineEventContext, sessions: ABSessionInfo[]): void {
-		// 	const context2 = context as any
-		// 	context2._knownSessions = sessions
-		// }
-		function createPieceInstance(
-			context: OnTimelineGenerateContext,
-			partInstanceId: PartInstanceId | string,
-			infiniteInstanceId?: PieceInstanceInfiniteId
-		): IBlueprintPieceInstance {
-			// This defines only the minimum required values for the method we are calling
-			const pieceInstance = {
-				_id: getRandomId(),
-				partInstanceId,
-				infinite: infiniteInstanceId ? { infiniteInstanceId } : undefined,
-			} as any
-
-			context.trackPieceInstances([pieceInstance])
-
-			return pieceInstance
-		}
-		function createTimelineObject(
-			partInstanceId: PartInstanceId | string | null,
-			infinitePieceInstanceId?: PieceInstanceInfiniteId,
-			isLookahead?: boolean
-		): OnGenerateTimelineObjExt {
-			// This defines only the minimum required values for the method we are calling
-			return {
-				partInstanceId,
-				infinitePieceInstanceId,
-				isLookahead: !!isLookahead,
-			} as any
-		}
-		function createPartInstance(id: string, partId: string, rank: number): DBPartInstance {
-			// This defines only the minimum required values for the method we are calling
-			return {
-				_id: id,
-				part: {
-					_id: partId,
-					_rank: rank,
-				},
-			} as any
-		}
-
-		test('getPieceABSessionId - knownSessions basic', async () => {
-			const { rundownId } = await setupDefaultRundownPlaylist(jobContext)
-			const rundown = (await jobContext.mockCollections.Rundowns.findOne(rundownId)) as DBRundown
-			expect(rundown).toBeTruthy()
-
-			// No sessions
-			{
-				const context = await getContext(rundown, undefined, undefined, undefined)
-				expect(context.knownSessions).toEqual([])
-			}
-
-			// some sessions
-			{
-				const sessions: ABSessionInfo[] = [{ id: 'abc', name: 'no' }]
-				// Mod the sessions to be returned by knownSessions
-				const moddedSessions = sessions.map((s) => ({ ...s, keep: true }))
-				await jobContext.mockCollections.RundownPlaylists.update(rundown.playlistId, {
-					$set: {
-						trackedAbSessions: moddedSessions,
-					},
-				})
-				const context = await getContext(rundown, undefined, undefined, undefined)
-				expect(context.knownSessions).toEqual(sessions)
-			}
-		})
-
-		test('getPieceABSessionId - bad parameters', async () => {
-			const { rundownId } = await setupDefaultRundownPlaylist(jobContext)
-			const rundown = (await jobContext.mockCollections.Rundowns.findOne(rundownId)) as DBRundown
-			expect(rundown).toBeTruthy()
-
-			{
-				const context = await getContext(rundown, undefined, undefined, undefined)
-
-				const piece1 = createPieceInstance(context, undefined as any)
-				expect(() => context.getPieceABSessionId(piece1, 'name0')).toThrow(
-					'Missing partInstanceId in call to getPieceABSessionId'
-				)
-
-				const piece2 = createPieceInstance(context, 'defdef')
-				expect(() => context.getPieceABSessionId(piece2, 'name0')).toThrow(
-					'Unknown partInstanceId in call to getPieceABSessionId'
-				)
-			}
-
-			{
-				const tmpPartInstance = createPartInstance('abcdef', 'aaa', 1)
-				const context = await getContext(rundown, undefined, undefined, tmpPartInstance)
-
-				const piece0 = createPieceInstance(context, 'defdef')
-				expect(() => context.getPieceABSessionId(piece0, 'name0')).toThrow(
-					'Unknown partInstanceId in call to getPieceABSessionId'
-				)
-
-				const piece1 = createPieceInstance(context, 'abcdef')
-				expect(context.getPieceABSessionId(piece1, 'name0')).toBeTruthy()
-			}
-		})
-
-		test('getPieceABSessionId - normal session', async () => {
-			const { rundownId } = await setupDefaultRundownPlaylist(jobContext)
-			const rundown = (await jobContext.mockCollections.Rundowns.findOne(rundownId)) as DBRundown
-			expect(rundown).toBeTruthy()
-
-			const nextPartInstance = createPartInstance('abcdef', 'aaa', 1)
-			const currentPartInstance = createPartInstance('12345', 'bbb', 0)
-			const context = await getContext(rundown, undefined, currentPartInstance, nextPartInstance)
-
-			// Get the id
-			const piece0 = createPieceInstance(context, nextPartInstance._id)
-			const expectedSessions: ABSessionInfo[] = [
-				{
-					id: getSessionId(0),
-					infiniteInstanceId: undefined,
-					name: 'name0',
-					partInstanceIds: [nextPartInstance._id],
-				},
-			]
-			expect(context.getPieceABSessionId(piece0, 'name0')).toEqual(expectedSessions[0].id)
-			expect(getAllKnownSessions(context)).toEqual(expectedSessions)
-			expect(context.knownSessions).toHaveLength(1)
-
-			// Should get the same id again
-			expect(context.getPieceABSessionId(piece0, 'name0')).toEqual(expectedSessions[0].id)
-			expect(getAllKnownSessions(context)).toEqual(expectedSessions)
-			expect(context.knownSessions).toHaveLength(1)
-
-			const piece1 = createPieceInstance(context, nextPartInstance._id)
-			expect(context.getPieceABSessionId(piece1, 'name0')).toEqual(expectedSessions[0].id)
-			expect(getAllKnownSessions(context)).toEqual(expectedSessions)
-			expect(context.knownSessions).toHaveLength(1)
-
-			// Try for the other part
-			const piece2 = createPieceInstance(context, currentPartInstance._id)
-			expect(context.getPieceABSessionId(piece2, 'name0')).not.toEqual(expectedSessions[0].id)
-			expect(context.knownSessions).toHaveLength(2)
-
-			// Or another name
-			expect(context.getPieceABSessionId(piece1, 'name1')).not.toEqual(expectedSessions[0].id)
-			expect(context.knownSessions).toHaveLength(3)
-		})
-
-		test('getPieceABSessionId - existing normal sessions', async () => {
-			const { rundownId } = await setupDefaultRundownPlaylist(jobContext)
-			const rundown = (await jobContext.mockCollections.Rundowns.findOne(rundownId)) as DBRundown
-			expect(rundown).toBeTruthy()
-
-			const nextPartInstance = createPartInstance('abcdef', 'aaa', 1)
-			const currentPartInstance = createPartInstance('12345', 'bbb', 0)
-
-			const expectedSessions: ABSessionInfo[] = [
-				{
-					id: 'current0',
-					name: 'name0',
-					partInstanceIds: [currentPartInstance._id],
-				},
-				{
-					id: 'current1',
-					name: 'name1',
-					partInstanceIds: [currentPartInstance._id],
-				},
-				{
-					id: 'next0',
-					name: 'name0',
-					partInstanceIds: [nextPartInstance._id],
-				},
-			]
-			await jobContext.mockCollections.RundownPlaylists.update(rundown.playlistId, {
-				$set: {
-					trackedAbSessions: expectedSessions,
-				},
-			})
-
-			const context = await getContext(rundown, undefined, currentPartInstance, nextPartInstance)
-
-			// Reuse the ids
-			const piece0 = createPieceInstance(context, currentPartInstance._id)
-			expect(context.getPieceABSessionId(piece0, 'name0')).toEqual(expectedSessions[0].id)
-			expect(getAllKnownSessions(context)).toEqual(expectedSessions)
-			expect(context.knownSessions).toHaveLength(1)
-
-			const piece1 = createPieceInstance(context, currentPartInstance._id)
-			expect(context.getPieceABSessionId(piece1, 'name1')).toEqual(expectedSessions[1].id)
-			expect(getAllKnownSessions(context)).toEqual(expectedSessions)
-			expect(context.knownSessions).toHaveLength(2)
-
-			const piece2 = createPieceInstance(context, nextPartInstance._id)
-			expect(context.getPieceABSessionId(piece2, 'name0')).toEqual(expectedSessions[2].id)
-			expect(getAllKnownSessions(context)).toEqual(expectedSessions)
-			expect(context.knownSessions).toHaveLength(3)
-		})
-
-		test('getPieceABSessionId - continue normal session from previous part', async () => {
-			const { rundownId } = await setupDefaultRundownPlaylist(jobContext)
-			const rundown = (await jobContext.mockCollections.Rundowns.findOne(rundownId)) as DBRundown
-			expect(rundown).toBeTruthy()
-
-			const nextPartInstance = createPartInstance('abcdef', 'aaa', 1)
-			const currentPartInstance = createPartInstance('12345', 'bbb', 0)
-
-			const context = await getContext(rundown, undefined, currentPartInstance, nextPartInstance)
-
-			const sessionId = getSessionId(0)
-			const piece0 = createPieceInstance(context, currentPartInstance._id)
-			expect(context.getPieceABSessionId(piece0, 'name0')).toEqual(sessionId)
-			expect(context.knownSessions).toHaveLength(1)
-
-			const piece2 = createPieceInstance(context, nextPartInstance._id)
-			expect(context.getPieceABSessionId(piece2, 'name0')).toEqual(sessionId)
-			expect(context.knownSessions).toHaveLength(1)
-		})
-
-		test('getPieceABSessionId - promote lookahead session from previous part', async () => {
-			const { rundownId } = await setupDefaultRundownPlaylist(jobContext)
-			const rundown = (await jobContext.mockCollections.Rundowns.findOne(rundownId)) as DBRundown
-			expect(rundown).toBeTruthy()
-
-			const previousPartInstance = createPartInstance('abcdef', 'aaa', 0)
-			const currentPartInstance = createPartInstance('12345', 'bbb', 1)
-			const distantPartId: PartId = protectString('future-part')
-
-			const lookaheadSessions: ABSessionInfo[] = [
-				{
-					id: 'lookahead0',
-					name: 'name0',
-					lookaheadForPartId: currentPartInstance.part._id,
-					partInstanceIds: [currentPartInstance._id],
-				},
-				{
-					id: 'lookahead1',
-					name: 'name1',
-					lookaheadForPartId: currentPartInstance.part._id,
-					partInstanceIds: undefined,
-				},
-				{
-					id: 'lookahead2',
-					name: 'name2',
-					lookaheadForPartId: distantPartId,
-					partInstanceIds: undefined,
-				},
-			]
-			await jobContext.mockCollections.RundownPlaylists.update(rundown.playlistId, {
-				$set: {
-					trackedAbSessions: lookaheadSessions,
-				},
-			})
-
-			const context = await getContext(rundown, previousPartInstance, currentPartInstance, undefined)
-
-			// lookahead0 is for us
-			const piece0 = createPieceInstance(context, currentPartInstance._id)
-			expect(context.getPieceABSessionId(piece0, 'name0')).toEqual('lookahead0')
-			expect(context.knownSessions).toHaveLength(1)
-
-			// lookahead1 is for us
-			const piece1 = createPieceInstance(context, currentPartInstance._id)
-			expect(context.getPieceABSessionId(piece1, 'name1')).toEqual('lookahead1')
-			expect(context.knownSessions).toHaveLength(2)
-
-			// lookahead2 is not for us, so we shouldnt get it
-			const sessionId = getSessionId(0)
-			const piece2 = createPieceInstance(context, currentPartInstance._id)
-			expect(context.getPieceABSessionId(piece2, 'name2')).toEqual(sessionId)
-			expect(context.knownSessions).toHaveLength(3)
-		})
-
-		test('getPieceABSessionId - infinite sessions', async () => {
-			const { rundownId } = await setupDefaultRundownPlaylist(jobContext)
-			const rundown = (await jobContext.mockCollections.Rundowns.findOne(rundownId)) as DBRundown
-			expect(rundown).toBeTruthy()
-
-			const nextPartInstance = createPartInstance('abcdef', 'aaa', 1)
-			const currentPartInstance = createPartInstance('12345', 'bbb', 10)
-
-			const context = await getContext(rundown, undefined, currentPartInstance, nextPartInstance)
-
-			// Start a new infinite session
-			const sessionId0 = getSessionId(0)
-			const infinite0 = protectString('infinite0')
-			const piece0 = createPieceInstance(context, currentPartInstance._id, infinite0)
-			expect(context.getPieceABSessionId(piece0, 'name0')).toEqual(sessionId0)
-			expect(context.knownSessions).toHaveLength(1)
-
-			// Double check the reuslt
-			expect(context.getPieceABSessionId(piece0, 'name0')).toEqual(sessionId0)
-			expect(context.knownSessions).toHaveLength(1)
-
-			// Normal piece in the same part gets different id
-			const sessionId1 = getSessionId(1)
-			const piece1 = createPieceInstance(context, currentPartInstance._id)
-			expect(context.getPieceABSessionId(piece1, 'name0')).toEqual(sessionId1)
-			expect(context.knownSessions).toHaveLength(2)
-
-			// Span session to a part with a lower rank
-			const piece2 = createPieceInstance(context, nextPartInstance._id, infinite0)
-			expect(context.getPieceABSessionId(piece2, 'name0')).toEqual(sessionId0)
-			expect(context.knownSessions).toHaveLength(2)
-		})
-
-		test('getTimelineObjectAbSessionId - bad parameters', async () => {
-			const { rundownId } = await setupDefaultRundownPlaylist(jobContext)
-			const rundown = (await jobContext.mockCollections.Rundowns.findOne(rundownId)) as DBRundown
-			expect(rundown).toBeTruthy()
-
-			const context = await getContext(rundown, undefined, undefined, undefined)
-
-			// no session needed
-			expect(context.getTimelineObjectAbSessionId({} as any, 'name0')).toBeUndefined()
-
-			// unknown partInstance
-			const obj1 = createTimelineObject('abcd')
-			expect(context.getTimelineObjectAbSessionId(obj1, 'name0')).toBeUndefined()
-		})
-
-		function generateGetTimelineObjectAbSessionIdSessions(
-			currentPartInstance: DBPartInstance,
-			nextPartInstance: DBPartInstance,
-			distantPartId: PartId,
-			infinite0: PieceInstanceInfiniteId,
-			infinite1: PieceInstanceInfiniteId
-		): ABSessionInfo[] {
-			return [
-				{
-					id: 'current0',
-					name: 'name0',
-					partInstanceIds: [currentPartInstance._id],
-				},
-				{
-					id: 'current1',
-					name: 'name1',
-					partInstanceIds: [currentPartInstance._id],
-				},
-				{
-					id: 'next0',
-					name: 'name0',
-					partInstanceIds: [nextPartInstance._id],
-				},
-				{
-					id: 'lookahead0',
-					name: 'name0',
-					lookaheadForPartId: currentPartInstance.part._id,
-					partInstanceIds: [currentPartInstance._id],
-				},
-				{
-					id: 'lookahead1',
-					name: 'name1',
-					lookaheadForPartId: currentPartInstance.part._id,
-					partInstanceIds: undefined,
-				},
-				{
-					id: 'lookahead2',
-					name: 'name2',
-					lookaheadForPartId: distantPartId,
-					partInstanceIds: undefined,
-				},
-				{
-					id: 'inf0',
-					name: 'name0',
-					infiniteInstanceId: infinite0,
-				},
-				{
-					id: 'inf1',
-					name: 'name0',
-					infiniteInstanceId: infinite1,
-				},
-			]
-		}
-
-		test('getTimelineObjectAbSessionId - normal', async () => {
-			const { rundownId } = await setupDefaultRundownPlaylist(jobContext)
-			const rundown = (await jobContext.mockCollections.Rundowns.findOne(rundownId)) as DBRundown
-			expect(rundown).toBeTruthy()
-
-			const nextPartInstance = createPartInstance('abcdef', 'aaa', 1)
-			const currentPartInstance = createPartInstance('12345', 'bbb', 10)
-
-			const existingSessions = generateGetTimelineObjectAbSessionIdSessions(
-				currentPartInstance,
-				createPartInstance('unknown', 'unknwon1', 9),
-				protectString('nope'),
-				protectString('infinite0'),
-				protectString('infinite1')
-			)
-			await jobContext.mockCollections.RundownPlaylists.update(rundown.playlistId, {
-				$set: {
-					trackedAbSessions: existingSessions,
-				},
-			})
-
-			const context = await getContext(rundown, undefined, currentPartInstance, nextPartInstance)
-
-			// no session recorded for partInstance
-			const obj1 = createTimelineObject(nextPartInstance._id)
-			expect(context.getTimelineObjectAbSessionId(obj1, 'name0')).toBeUndefined()
-
-			// partInstance with session
-			const obj2 = createTimelineObject(currentPartInstance._id)
-			expect(context.getTimelineObjectAbSessionId(obj2, 'name0')).toEqual('current0')
-			expect(context.getTimelineObjectAbSessionId(obj2, 'name1')).toEqual('current1')
-
-			// // define a session now
-			// overwriteKnownSessions(context, [{
-			// 	{
-			// 		id: 'current0',
-			// 		name: 'name0',
-			// 		partInstanceIds: [currentPartInstance._id],
-			// 	},
-			// }])
-
-			// Ensure the sessions havent changed
-			expect(getAllKnownSessions(context)).toEqual(existingSessions)
-		})
-
-		test('getTimelineObjectAbSessionId - lookahead', async () => {
-			const { rundownId } = await setupDefaultRundownPlaylist(jobContext)
-			const rundown = (await jobContext.mockCollections.Rundowns.findOne(rundownId)) as DBRundown
-			expect(rundown).toBeTruthy()
-
-			const nextPartInstance = createPartInstance('abcdef', 'aaa', 1)
-			const currentPartInstance = createPartInstance('12345', 'bbb', 10)
-
-			const distantPartId: PartId = protectString('distant0')
-			const existingSessions = generateGetTimelineObjectAbSessionIdSessions(
-				currentPartInstance,
-				nextPartInstance,
-				distantPartId,
-				protectString('infinite0'),
-				protectString('infinite1')
-			)
-			await jobContext.mockCollections.RundownPlaylists.update(rundown.playlistId, {
-				$set: {
-					trackedAbSessions: [...existingSessions],
-				},
-			})
-
-			const context = await getContext(rundown, undefined, currentPartInstance, nextPartInstance)
-
-			// no session if no partId
-			const obj1 = createTimelineObject(null, undefined, true)
-			expect(context.getTimelineObjectAbSessionId(obj1, 'name0')).toBeUndefined()
-			expect(context.knownSessions).toHaveLength(0)
-
-			// existing 'distant' lookahead session
-			const obj2 = createTimelineObject(unprotectString(distantPartId), undefined, true)
-			expect(context.getTimelineObjectAbSessionId(obj2, 'name2')).toEqual('lookahead2')
-			expect(context.knownSessions).toHaveLength(1)
-
-			// new 'distant' lookahead session
-			const obj2a = createTimelineObject(unprotectString(distantPartId), undefined, true)
-			expect(context.getTimelineObjectAbSessionId(obj2a, 'name0')).toEqual(getSessionId(0))
-			expect(context.knownSessions).toHaveLength(2)
-			existingSessions.push({
-				id: getSessionId(0),
-				lookaheadForPartId: distantPartId,
-				name: 'name0',
-			})
-
-			// current partInstance session
-			const obj3 = createTimelineObject(currentPartInstance._id, undefined, true)
-			expect(context.getTimelineObjectAbSessionId(obj3, 'name1')).toEqual('current1')
-			expect(context.knownSessions).toHaveLength(3)
-
-			// next partInstance session
-			const obj4 = createTimelineObject(nextPartInstance._id, undefined, true)
-			expect(context.getTimelineObjectAbSessionId(obj4, 'name0')).toEqual('next0')
-			expect(context.knownSessions).toHaveLength(4)
-
-			// next partInstance new session
-			const obj5 = createTimelineObject(nextPartInstance._id, undefined, true)
-			expect(context.getTimelineObjectAbSessionId(obj5, 'name1')).toEqual(getSessionId(1))
-			expect(context.knownSessions).toHaveLength(5)
-
-			existingSessions.push({
-				id: getSessionId(1),
-				lookaheadForPartId: nextPartInstance.part._id,
-				name: 'name1',
-				partInstanceIds: [nextPartInstance._id],
-			})
-
-			// Ensure the sessions havent changed
-			expect(getAllKnownSessions(context)).toEqual(existingSessions)
-		})
-
-		test('getTimelineObjectAbSessionId - lookahead2', async () => {
-			const { rundownId } = await setupDefaultRundownPlaylist(jobContext)
-			const rundown = (await jobContext.mockCollections.Rundowns.findOne(rundownId)) as DBRundown
-			expect(rundown).toBeTruthy()
-
-			const nextPartInstance = createPartInstance('abcdef', 'aaa', 1)
-			const currentPartInstance = createPartInstance('12345', 'bbb', 10)
-
-			const distantPartId: PartId = protectString('distant0')
-			const infinite0: PieceInstanceInfiniteId = protectString('infinite0')
-			const infinite1: PieceInstanceInfiniteId = protectString('infinite1')
-			const existingSessions = generateGetTimelineObjectAbSessionIdSessions(
-				currentPartInstance,
-				nextPartInstance,
-				distantPartId,
-				infinite0,
-				infinite1
-			)
-			await jobContext.mockCollections.RundownPlaylists.update(rundown.playlistId, {
-				$set: {
-					trackedAbSessions: [...existingSessions],
-				},
-			})
-
-			const context = await getContext(rundown, undefined, currentPartInstance, nextPartInstance)
-
-			const obj1 = createTimelineObject(currentPartInstance._id, infinite0)
-			expect(context.getTimelineObjectAbSessionId(obj1, 'name0')).toEqual('inf0')
-			expect(context.knownSessions).toHaveLength(1)
-
-			const obj2 = createTimelineObject(null, infinite1)
-			expect(context.getTimelineObjectAbSessionId(obj2, 'name0')).toEqual('inf1')
-			expect(context.knownSessions).toHaveLength(2)
-
-			const obj3 = createTimelineObject(null, protectString('fake'))
-			expect(context.getTimelineObjectAbSessionId(obj3, 'name0')).toBeUndefined()
-			expect(context.knownSessions).toHaveLength(2)
-
-			// Ensure the sessions havent changed
-			expect(getAllKnownSessions(context)).toEqual(existingSessions)
 		})
 	})
 })
